XML course project
Develop a Java Windows application, which receives and saves students’
feedbacks on courses to an XML file. The application should consist of a
text box for student name, a combo box for a list of courses and a combo
box for available grades (1-5). After filling the fields and pressing
Submit button, the application should send data to the XML file. After
saving data to the XML file, the application should display summary of
students' feedbacks